requirements.txt
praw>=7.7,<8.0
twikit>=2.1
openai>=1.30
python-telegram-bot>=21.0
telethon>=1.35
pyyaml>=6.0
aiosqlite>=0.20
apscheduler>=3.10,<4.0
click>=8.1
requests>=2.31
playwright>=1.40
rich>=13.0
fastapi>=0.110
uvicorn[standard]>=0.27
